+++ src/components/Tables/TableThree.tsx
@@ -1,12 +1,11 @@
 'use client'
 
-// import { Branches } from "@/types/branches";
 import Image from "next/image";
 import Link from "next/link";
 import { SetStateAction, useEffect, useState } from "react";
 
 type Doctor = {
-  id: number;
+  _id: number;
   image: string;
   name: string;
   position: string; 
@@ -16,6 +15,12 @@ interface TableProps {
   limit?: number | null; 
   showPagination?: boolean; 
 }
+
+type DoctorsResponse = {
+  doctors: Doctor[];
+  currentPage: number;
+  totalPages: number;
+};
 
 const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true }) => {
     const [doctors, setDoctors] = useState<Doctor[]>([]);
@@ -27,43 +32,45 @@ const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true 
     const [loadingDelete, setLoadingDelete] = useState(false);
    
     const itemsPerPage = 15;
-  
-    useEffect(() => {
-      const fetchDoctors = async (currentPage: number) => {
-        try {
-          const response = await fetch(`/api/doctors?page=${currentPage}`);
-    
-          if (!response.ok) {
-            throw new Error(`HTTP error! Status: ${response.status}`);
-          }
-    
-          const result = await response.json(); 
-          const data = limit ? result.data.slice(0, limit) : result.data;
-          
-          setDoctors(data); // Ambil hanya bagian 'data'
-          setCurrentPage(result.pagination.currentPage);
-          setTotalPages(result.pagination.totalPages);
-        } catch (error) {
-          console.error("Error fetching doctors:", error);
-        } finally {
-          setLoading(false);
+
+    const fetchDoctors = async (page = 1, limit?: number | null) => {
+      try {
+        const response = await fetch(`/api/doctors?page=${page}`);
+        if (!response.ok) {
+          throw new Error("Gagal mengambil data dokter");
         }
-      };
+    
+        const result: DoctorsResponse = await response.json();
+        
+        // Jika ada limit, potong hasil data dokter
+        const doctorsData = limit ? result.doctors.slice(0, limit) : result.doctors;
+    
+        
+        setDoctors(doctorsData);
+        setCurrentPage(result.currentPage);
+        setTotalPages(result.totalPages);
+      } catch (error) {
+        console.error("Gagal mengambil data dokter:", error);
+      }
+    };
+    
+    // Ambil halaman pertama saat load
+    useEffect(() => {
+      fetchDoctors(1, limit);
+    }, [limit]);
 
-      fetchDoctors(currentPage);
-    }, [currentPage, limit]);
-
-    const handleDeleteDoctor = async (id: string | number) => {
+    const handleDeleteDoctor = async (_id: string | number) => {
+      if (!selectedDoctor) return;
       try {
         setLoadingDelete(true);
-        const response = await fetch(`/api/doctorsDelete/${id}`, {
+        const response = await fetch(`api/doctors/${selectedDoctor._id}`, {
           method: 'DELETE',
         });
     
         if (!response.ok) {
           throw new Error(response.statusText);
         }
-        setDoctors((prevDoctors) => prevDoctors.filter((doctor) => doctor.id !== id));
+        setDoctors((prevDoctors) => prevDoctors.filter((doctor) => doctor._id !== _id));
         setSelectedDoctor(null);
         setIsOpen(false);
       } catch (error) {
@@ -76,9 +83,9 @@ const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true 
   return (
     <div className="rounded-[10px] border border-stroke bg-white p-4 shadow-1 dark:border-dark-3 dark:bg-gray-dark dark:shadow-card sm:p-7.5">
       <div className="max-w-full overflow-x-auto">
-      {loading ? (
+      {/* {loading ? (
         <p className="text-center text-gray-500 dark:text-white mb-5 text-2xl font-semibold">Loading...</p>
-      ) : (
+      ) : ( */}
         <table className="w-full table-auto">
           <thead>
             <tr className="bg-[#F7F9FC] text-left dark:bg-dark-2">
@@ -100,7 +107,7 @@ const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true 
             </tr>
           </thead>
           <tbody>
-            {doctors.map((doctor, index) => (
+            {doctors.map((doctor: any, index) => (
               <tr key={index}>
                 <td
                   className={`border-[#eee] px-4 text-center py-4 dark:border-dark-3 w-0 xl:pl-9 ${
@@ -138,7 +145,7 @@ const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true 
                   className={`border-[#eee] px-4 py-4 dark:border-dark-3 xl:pr-7.5 ${index === doctors.length - 1 ? "border-b-0" : "border-b"}`}
                 >
                   <div className="flex items-center justify-end space-x-3.5">
-                    <Link href={`/doctors/edit/${doctor.id}`} className="p-0 m-0 flex items-center justify-center">
+                    <Link href={`/doctors/edit/${doctor._id}`} className="p-0 m-0 flex items-center justify-center">
                       <button className="hover:text-orange-400">
                         <svg xmlns="http://www.w3.org/2000/svg" width="22" height="22" viewBox="0 0 24 24"><path fill="currentColor" fillRule="evenodd" d="M21.455 5.416a.75.75 0 0 1-.096.943l-9.193 9.192a.75.75 0 0 1-.34.195l-3.829 1a.75.75 0 0 1-.915-.915l1-3.828a.8.8 0 0 1 .161-.312L17.47 2.47a.75.75 0 0 1 1.06 0l2.829 2.828a1 1 0 0 1 .096.118m-1.687.412L18 4.061l-8.518 8.518l-.625 2.393l2.393-.625z" clipRule="evenodd"/><path fill="currentColor" d="M19.641 17.16a44.4 44.4 0 0 0 .261-7.04a.4.4 0 0 1 .117-.3l.984-.984a.198.198 0 0 1 .338.127a46 46 0 0 1-.21 8.372c-.236 2.022-1.86 3.607-3.873 3.832a47.8 47.8 0 0 1-10.516 0c-2.012-.225-3.637-1.81-3.873-3.832a46 46 0 0 1 0-10.67c.236-2.022 1.86-3.607 3.873-3.832a48 48 0 0 1 7.989-.213a.2.2 0 0 1 .128.34l-.993.992a.4.4 0 0 1-.297.117a46 46 0 0 0-6.66.255a2.89 2.89 0 0 0-2.55 2.516a44.4 44.4 0 0 0 0 10.32a2.89 2.89 0 0 0 2.55 2.516c3.355.375 6.827.375 10.183 0a2.89 2.89 0 0 0 2.55-2.516"/></svg>
                       </button>
@@ -161,7 +168,7 @@ const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true 
             ))}
           </tbody>
         </table>
-      )}
+      {/* )} */}
       {isOpen && selectedDoctor && (
         <div className="fixed top-0 left-0 w-full h-full bg-black bg-opacity-35 z-999 flex justify-center items-center z-50">
           <div className="bg-white rounded-2xl p-6 py-9 w-1/3 shadow-lg">
@@ -177,7 +184,7 @@ const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true 
               <button className="bg-gray-200 hover:bg-gray-300 text-lg text-gray-600 py-2 px-5 rounded-lg cursor-pointer" onClick={() => setSelectedDoctor(null)}>
                 Cancel
               </button>
-              <button className="bg-red-500 hover:bg-red-600 text-lg text-white py-2 px-5 rounded-lg cursor-pointer" onClick={() => handleDeleteDoctor(selectedDoctor.id)}>
+              <button className="bg-red-500 hover:bg-red-600 text-lg text-white py-2 px-5 rounded-lg cursor-pointer" onClick={() => handleDeleteDoctor(selectedDoctor._id)}>
                 {loadingDelete ? 'Deleting...' : 'Delete'}
               </button>
             </div>
@@ -188,8 +195,8 @@ const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true 
         <div className="flex justify-center mt-4 space-x-2">
             <button
               className={`px-4 py-2 rounded ${currentPage === 1 ? "bg-[#F7F9FC] dark:bg-dark-2 cursor-not-allowed" : "bg-orange-400 text-white hover:bg-orange-600"}`}
-              disabled={currentPage === 1}
-              onClick={() => setCurrentPage(currentPage - 1)}
+              disabled={currentPage === 1} 
+              onClick={() => fetchDoctors(currentPage - 1)}
             >
               Prev
             </button>
@@ -200,8 +207,8 @@ const TableThree: React.FC<TableProps> = ({ limit = null, showPagination = true 
 
             <button
               className={`px-4 py-2 rounded ${currentPage === totalPages ? "bg-[#F7F9FC] dark:bg-dark-2 cursor-not-allowed" : "bg-orange-400 text-white hover:bg-orange-600"}`}
-              disabled={currentPage === totalPages}
-              onClick={() => setCurrentPage(currentPage + 1)}
+              disabled={currentPage === totalPages} 
+              onClick={() => fetchDoctors(currentPage + 1)}
             >
               Next
             </button>
